+++ pyobjc-framework-CoreWLAN/PyObjCTest/test_corewlan.py
@@ -2,12 +2,6 @@ import CoreWLAN
 from PyObjCTools.TestSupport import TestCase
 
 
-class TestCoreWLAN(TestCase):
-    def test_constants(self):
-        self.assertEqual(CoreWLAN.CoreWLANFrameworkVersionNumber2_0, 200)
-        self.assertIsInstance(CoreWLAN.CoreWLANFrameworkVersionNumber, float)
-
-
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
         self.assertCallableMetadataIsSane(CoreWLAN)