+++ pyobjc-framework-PassKit/Lib/PassKit/_metadata.py
@@ -1177,7 +1177,7 @@ try:
     r(b"PKStoredValuePassBalance", b"isEqualToBalance:", {"retval": {"type": b"Z"}})
     r(b"PKStoredValuePassProperties", b"isBlacklisted", {"retval": {"type": b"Z"}})
     r(b"PKStoredValuePassProperties", b"isBlocked", {"retval": {"type": b"Z"}})
-    r(b"PKStoredValuePassProperties", b"isInStation", {"retval": {"type": b"Z"}})
+    r(b"PKStoredValuePassProperties", b"isInStation", {"retval": {"type": "Z"}})
     r(
         b"PKSuicaPassProperties",
         b"isBalanceAllowedForCommute",